now supports add/remove operations on the orig
we now listen to add/remove operations on the
original collection and react to it not by
resetting but by surgically adding/removing.

also modified behavior to not support the model
argument or model operations. basically this
collection now protects against direct 
modifications as it is meant to be a filter on
another collection always.

removing backbone-relational testing dependency
it served no purpose.